# api/openapi.yaml
openapi: 3.1.0
info:
  title: TransitGo Mobile BFF (Training)
  version: 0.1.0
  description: |
    Compact training API for TransitGo. External integrations are mocked.
    - FR-001..FR-004 Journey search (TRGO-20)
    - FR-007, FR-009 Departures & nearby stops (TRGO-18)
    - FR-020 Identity foundation (TRGO-14)
    - FR-022..FR-024 Payments PoC (TRGO-16)
servers:
  - url: http://localhost:3001
    description: Local dev (journey-planning)
  - url: http://localhost:3002
    description: Local dev (realtime)
  - url: http://localhost:3003
    description: Local dev (identity)
  - url: http://localhost:3004
    description: Local dev (payments)
  - url: http://localhost:3005
    description: Local dev (ticketing)
paths:
  /v1/journeys:
    get:
      summary: Journey search
      description: |
        Implements FR-001, FR-002, FR-003 (TRGO-20).
        Adapts to OpenTripPlanner (mocked in this training repo).
      parameters:
        - in: query
          name: origin
          required: true
          schema: { type: string }
        - in: query
          name: destination
          required: true
          schema: { type: string }
        - in: query
          name: time
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: arriveBy
          required: false
          schema: { type: boolean, default: false }
        - in: query
          name: accessibility
          required: false
          schema: { type: string, enum: [step_free, any], default: any }
        - in: query
          name: maxWalkMeters
          required: false
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Journey options
          content:
            application/json:
              schema:
                type: object
                properties:
                  options:
                    type: array
                    items:
                      $ref: '#/components/schemas/JourneyOption'
        "400":
          $ref: '#/components/responses/BadRequest'
  /v1/places/autocomplete:
    get:
      summary: Place autocomplete
      description: Implements TRGO-68 (contract stub for CAP-1).
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Matching places
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Place'
  /v1/departures/{stopId}:
    get:
      summary: Stop departures
      description: Implements FR-007 and FR-009 (TRGO-18). GTFS-RT is mocked.
      parameters:
        - in: path
          name: stopId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Departures at stop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartureBoard'
  /v1/nearby:
    get:
      summary: Nearby stops
      description: Implements FR-009 (TRGO-18). Mocked data.
      parameters:
        - in: query
          name: lat
          required: true
          schema: { type: number }
        - in: query
          name: lng
          required: true
          schema: { type: number }
        - in: query
          name: radius
          required: false
          schema: { type: integer, default: 300 }
      responses:
        "200":
          description: Nearby stops
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Place'
  /v1/identity/login:
    post:
      summary: Stub login
      description: Implements FR-020 (TRGO-14). Identity is mocked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
              required: [email, password]
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
  /v1/identity/.well-known/jwks.json:
    get:
      summary: JWKS
      description: Public signing keys (mock JWKS) to validate JWTs.
      responses:
        "200":
          description: JWKS
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/Jwk'
  /v1/identity/profile:
    get:
      summary: Profile
      description: Returns current user profile from JWT (mock).
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  sub: { type: string }
                  email: { type: string }
                  consents:
                    type: array
                    items: { $ref: '#/components/schemas/Consent' }
  /v1/identity/consents:
    put:
      summary: Save consents
      description: Stores user GDPR consents (mock).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                consents:
                  type: array
                  items: { $ref: '#/components/schemas/Consent' }
      responses:
        "204":
          description: Saved
  /v1/payments/intents:
    post:
      summary: Create a payment intent (mock)
      description: |
        Implements FR-022..FR-024 (TRGO-16). Uses a provider-agnostic adapter with a mock PSP.
        NOTE: No real PSP named due to RISK-004.
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: integer, minimum: 1, description: Minor units (e.g. cents) }
                currency: { type: string, enum: [EUR] }
              required: [amount, currency]
      responses:
        "200":
          description: Mock intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
  /v1/payments/3ds/simulate:
    post:
      summary: Simulate 3DS success
      description: Moves an intent to succeeded (training only).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                intentId: { type: string }
              required: [intentId]
      responses:
        "204":
          description: State changed
  /v1/payments/webhook:
    post:
      summary: PSP webhook (mock)
      description: Verifies HMAC signature.
      parameters:
        - in: header
          name: x-psp-signature
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          text/plain:
            schema: { type: string }
      responses:
        "200":
          description: OK
  /v1/products:
    get:
      summary: Ticket products catalogue
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Product' }
  /v1/orders:
    get:
      summary: Order history
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
    post:
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId: { type: string }
                amount: { type: integer }
                currency: { type: string, enum: [EUR] }
              required: [productId, amount, currency]
      responses:
        "201":
          description: Created
  /v1/refunds:
    get:
      summary: Refunds
      responses:
        "200":
          description: Refunds
    post:
      summary: Create refund
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orderId: { type: string }
                amount: { type: integer }
              required: [orderId, amount]
      responses:
        "201":
          description: Created
  /v1/tickets:
    get:
      summary: List tickets
      parameters:
        - in: query
          name: userId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tickets
    post:
      summary: Create ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId: { type: string }
                userId: { type: string }
              required: [productId, userId]
      responses:
        "201":
          description: Created
  /v1/tickets/activate:
    post:
      summary: Activate ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
              required: [id]
      responses:
        "204":
          description: Activated
  /v1/tickets/{id}/qr:
    get:
      summary: Get QR token
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Token
  /v1/tickets/validate:
    post:
      summary: Validate QR token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
              required: [token]
      responses:
        "200":
          description: Validation result
components:
  schemas:
    JourneyOption:
      type: object
      properties:
        summary: { type: string }
        durationMinutes: { type: integer }
        legs:
          type: array
          items:
            type: object
            properties:
              mode: { type: string, enum: [WALK, BUS, TRAM, TROLLEYBUS] }
              from: { type: string }
              to: { type: string }
              startTime: { type: string, format: date-time }
              endTime: { type: string, format: date-time }
    Place:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        lat: { type: number }
        lon: { type: number }
    DepartureBoard:
      type: object
      properties:
        stopId: { type: string }
        departures:
          type: array
          items:
            type: object
            properties:
              line: { type: string }
              destination: { type: string }
              plannedTime: { type: string, format: date-time }
              expectedInMinutes: { type: integer }
    PaymentIntent:
      type: object
      properties:
        id: { type: string }
        clientSecret: { type: string }
        amount: { type: integer }
        currency: { type: string }
        status: { type: string, enum: [requires_action, processing, succeeded] }
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        priceCents: { type: integer }
        durationHours: { type: integer }
    Order:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        amount: { type: integer }
        currency: { type: string }
        createdAt: { type: string, format: date-time }
        status: { type: string, enum: [paid, refunded] }
    Jwk:
      type: object
      properties:
        kty: { type: string }
        n: { type: string }
        e: { type: string }
        kid: { type: string }
        alg: { type: string }
        use: { type: string }
    Consent:
      type: object
      properties:
        id: { type: string }
        granted: { type: boolean }
        timestamp: { type: string, format: date-time }
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
                  details: { type: object }